# apps/api-server/src/chat/dto/create-chat.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { IsNumber } from "class-validator";

export class CreateChatDto {
  @ApiProperty()
  @IsNumber()
  invitorId: number;

  @ApiProperty()
  @IsNumber()
  inviteeId: number;
}