# apps/api-server/swagger.json
{"openapi":"3.0.0","paths":{"/auth/signup":{"post":{"operationId":"AuthController_signup","summary":"Create account and login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signin","summary":"Login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/loginUserDTO"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/refresh":{"get":{"operationId":"AuthController_refreshTokens","summary":"Refresh token","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"],"security":[{"acces-token":[]}]}},"/users":{"get":{"operationId":"UsersController_getUsers","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"404":{"description":""}},"tags":["users"]},"post":{"operationId":"UsersController_createUser","summary":"Create a user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/users/{id}":{"get":{"operationId":"UsersController_getUserById","summary":"Get a user depending on passed id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"patch":{"operationId":"UsersController_updateUser","summary":"Update a user depending on passed id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatchUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]},"delete":{"operationId":"UsersController_deleteUser","summary":"Delete a user depending on passed id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/users/doctor":{"post":{"operationId":"UsersController_createDoctor","summary":"Create a doctor","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/chat-event/{chatId}":{"get":{"operationId":"ChatEventController_getChatEventsFromChatId","parameters":[{"name":"chatId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["ChatEvent"]},"post":{"operationId":"ChatEventController_createChatEvent","summary":"Get a chat-event depending on passed chat id","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChatEventDto"}}}},"responses":{"201":{"description":""}},"tags":["ChatEvent"]}},"/chat/{id}":{"get":{"operationId":"ChatController_getChat","summary":"Get a chat depending on passed id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}}},"tags":["Chat"]},"delete":{"operationId":"ChatController_deleteChat","summary":"Delete a chat depending on passed id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}}},"tags":["Chat"]}},"/chat":{"post":{"operationId":"ChatController_createChat","summary":"Create a chat depending on passed id","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChatDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}}},"tags":["Chat"]}},"/chat/user/{id}":{"get":{"operationId":"ChatController_getChatByUserId","summary":"Get chat(s) where an user appear","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Chat"}}}}}},"tags":["Chat"]}},"/ai":{"post":{"operationId":"AiController_sendMessage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAiDto"}}}},"responses":{"201":{"description":""}},"tags":["AI"]}},"/appointment":{"post":{"operationId":"AppointmentController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAppointmentDto"}}}},"responses":{"201":{"description":""}},"tags":["Appointments"]}},"/appointment/{id}":{"get":{"operationId":"AppointmentController_findOneAppointment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Appointments"]}},"/appointment/user":{"get":{"operationId":"AppointmentController_findAllForOneUser","summary":"Get all appointment for the user logged","parameters":[],"responses":{"200":{"description":""}},"tags":["Appointments"],"security":[{"acces-token":[]}]}},"/appointment/accept/{appointmentId}":{"patch":{"operationId":"AppointmentController_acceptAppointment","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Appointments"],"security":[{"acces-token":[]}]}},"/appointment/decline/{appointmentId}":{"patch":{"operationId":"AppointmentController_declineAppointment","parameters":[{"name":"appointmentId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Appointments"],"security":[{"acces-token":[]}]}}},"info":{"title":"Web-Services API","description":"API du cours de web-services","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"acces-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"heightInCm":{"type":"number"},"weightInKg":{"type":"number"},"ssn":{"type":"string"},"proDoctorNumber":{"type":"string"}},"required":["email","password","firstName","lastName","phone","heightInCm","weightInKg","ssn","proDoctorNumber"]},"loginUserDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"User":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"heightInCm":{"type":"number"},"weightInKg":{"type":"number"},"ssn":{"type":"string"},"proDoctorNumber":{"type":"string"}},"required":["email","password","firstName","lastName","phone","heightInCm","weightInKg","ssn","proDoctorNumber"]},"PatchUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"heightInCm":{"type":"number"},"weightInKg":{"type":"number"},"ssn":{"type":"string"},"proDoctorNumber":{"type":"string"}},"required":["email","password","firstName","lastName","phone","heightInCm","weightInKg","ssn","proDoctorNumber"]},"CreateChatEventDto":{"type":"object","properties":{"chatId":{"type":"number"},"userId":{"type":"number"},"data":{"type":"string"},"chatEventType":{"type":"number"}},"required":["chatId","userId","data","chatEventType"]},"Chat":{"type":"object","properties":{"invitorId":{"type":"number"},"inviteeId":{"type":"number"}},"required":["invitorId","inviteeId"]},"CreateChatDto":{"type":"object","properties":{"invitorId":{"type":"number"},"inviteeId":{"type":"number"}},"required":["invitorId","inviteeId"]},"CreateAiDto":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"CreateAppointmentDto":{"type":"object","properties":{"patientId":{"type":"number"},"doctorId":{"type":"number"},"location":{"type":"string"},"date":{"format":"date-time","type":"string"}},"required":["patientId","doctorId","location","date"]}}}}